changed some saving methods
