+++ frontend/set-up/pages/age.html
@@ -77,7 +77,6 @@
         </div>
 
     </div>   
-     <script type = "module" src="../scripts/universalPicker.js"></script>
      <script src="../scripts/script-set-up.js"></script>
 </body>
 </html>
