+++ frontend/set-up/pages/height.html
@@ -69,7 +69,6 @@
         </div>
 
     </div>   
-     <script type = "module" src="../scripts/universalPicker.js"></script>
      <script src="../scripts/script-set-up.js"></script>
 </body>
 </html>
