+++ frontend/set-up/pages/weight.html
@@ -68,7 +68,6 @@
         </div>
 
     </div>   
-     <script type = "module" src="../scripts/universalPicker.js"></script>
      <script src="../scripts/script-set-up.js"></script>
 </body>
 </html>
