+++ frontend/set-up/style-sheets/log-in.css
@@ -21,7 +21,7 @@
     width: 300px;
     height: 50px;
     border-radius: 50px;
-    gap: 20px;
+    gap: 10px;
     background-color: #1F2937;
 }
 
@@ -105,7 +105,11 @@
 
   background-color: #1F2937;
   color: #FFFFFF;
-  
+}
+
+#email-logo{
+    width: 30px;
+    height: 25px;
 }
